requirements.txt
python==3.10
PyQt5==5.15.9
OpenCv-python
numpy==1.24.3
pytz
requests
PyQtWebEngine
pillow
python-dotenv
tensorflow==2.10.1
matplotlib
tqdm
torch
transformers
json
os
sqlite3
configparser
scikit-learn
cartopy
folium
